-adding 4 moving averages as tools -adding class to manage moving averages -different kind of calculating MACD -lack of intersect checking

+++ src/main/java/Indexes/IStockIndex.java
@@ -3,5 +3,5 @@ package Indexes;
 
 public interface IStockIndex
 {
-    public double calculate();
+    public int calculate();
 }
